Hotel.checkin() method was created.
Folling steps are taken to create a checkin() method.
confirmationNumber was looked in the hashmap called bookingByConfirmationNumber.
If key matches confirmation number then the Booking object is extracted from Map.
Room id is extracted from the Booking object and the values are stored in a hashmap that take roomId as key and Booking object as value called activeBookingsByRoomId.
The status of booking was changed by calling booking.checkIn().
If confirmationNumber does not match then a RuntimeException is thrown with message regarding "Booking not found." error message.