BookingCTL.creditDetailsEntered() method is created.
BookingCTL.creditDetailsEntered() method is created using following logic:
- arguments provided for method is used to check whether the creditcard can be authorised using CreditAuthorizer.authorize() method.
If it return false then it throw RuntimeException.
If it return true then the state is changed to approved and a new Booking object is created.
After creating an object of Booking, its method is used to generate data and is passed to bookingUI.displayConfirmedBooking() which shows the message to Guest.
After that, the state of the bookingUI is set to COMPLETED and so to the state of BookingCTL.